+++ backend/config/routes.rb
@@ -3,35 +3,32 @@ Rails.application.routes.draw do
   post 'drink/:user_id', to: 'drink#create'
   
   # drinkコントローラのshowアクションへのルーティング
-  # 例: GET /drink/1
   get 'drink/:user_id', to: 'drink#show'
 
   # drinkコントローラのshowアクションへのルーティング
-  # 例: GET /custom/1
-  get 'custom/:user_id', to: 'drink#showCustom'  
+  get 'custom/:user_id', to: 'customs#showCustom'  
 
   # DrinksControllerに対する新しいルートを追加
-  post 'drinks/update_drink_result/:user_id', to: 'drink#update_drink_result'
+  post 'drinks/update_drink_result/:user_id', to: 'customs#update_drink_result'
 
   # 新しいAPIエンドポイントのルート設定
   get 'drinks/history/:user_id', to: 'drink#drink_history'  
 
   # 履歴画面からカスタム画面に飛ぶときにカスタムをDBから取得する際
-  # get 'drinks/history/customs/:id', to: 'drink#showCustomFromHistory'  
-  get 'drinks/:id', to: 'drink#showCustomFromHistory'  
+  get 'drinks/:id', to: 'customs#showCustomFromHistory'  
 
   # ドリンクのカスタム情報を更新するためのルーティング
-  post 'drinks/update/:id', to: 'drink#update_custom'
+  post 'drinks/update/:id', to: 'customs#update_custom'
 
   # ドリンクのコメントを編集するためのルーティング
-  get 'comment/:id', to: 'drink#show_comment'
+  get 'comment/:id', to: 'comment#show_comment'
 
   # ドリンクのコメントを更新するためのルーティング
-  post 'update_comment/:id', to: 'drink#update_comment'
+  post 'update_comment/:id', to: 'comment#update_comment'
 
   # サインアップ時のルーティング
-  post '/signup', to: 'users#create'
+  post 'signup', to: 'users#create'
 
   # メールアドレスからユーザIDを取得するためのルーティング
-  get '/find_user_id', to: 'users#find_user_id'  
+  get 'find_user_id', to: 'users#find_user_id'  
 end
